+++ src/main/kotlin/di/RepositoryModule.kt
@@ -1,12 +1,14 @@
 package com.ua.astrumon.di
 
 import com.ua.astrumon.data.db.repository.BirthdayGreetingRepositoryImpl
+import com.ua.astrumon.data.db.repository.BotMetaRepositoryImpl
 import com.ua.astrumon.data.db.repository.ChatRepositoryImpl
 import com.ua.astrumon.data.db.repository.GroupMemberRepositoryImpl
 import com.ua.astrumon.data.db.repository.GroupRepositoryImpl
 import com.ua.astrumon.data.db.repository.MemberChatRepositoryImpl
 import com.ua.astrumon.data.db.repository.MemberRepositoryImpl
 import com.ua.astrumon.domain.repository.BirthdayGreetingRepository
+import com.ua.astrumon.domain.repository.BotMetaRepository
 import com.ua.astrumon.domain.repository.ChatRepository
 import com.ua.astrumon.domain.repository.GroupMemberRepository
 import com.ua.astrumon.domain.repository.GroupRepository
@@ -21,4 +23,5 @@ val repositoryModule = module {
     single<GroupMemberRepository> { GroupMemberRepositoryImpl() }
     single<ChatRepository> { ChatRepositoryImpl() }
     single<BirthdayGreetingRepository> { BirthdayGreetingRepositoryImpl() }
+    single<BotMetaRepository> { BotMetaRepositoryImpl() }
 }